requirements.txt
flask
python-dateutil
psycopg2-binary
python-dotenv